requirements.txt
praw==7.8.1
requests==2.32.3
google-generativeai==0.7.2
python-dotenv==1.0.1
httpx==0.27.0